test/rtpw: Cast Sleep argument on MSVC
test/rtpw.c(561): warning C4244: 'function': conversion from 'double' to 'DWORD', possible loss of data

+++ test/rtpw.c
@@ -96,7 +96,7 @@
 
 #ifndef HAVE_USLEEP
 #ifdef HAVE_WINDOWS_H
-#define usleep(us) Sleep((us) / 1000)
+#define usleep(us) Sleep(((DWORD)us) / 1000)
 #else
 #define usleep(us) sleep((us) / 1000000)
 #endif
